api/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-jose[cryptography]
bcrypt
python-multipart
